+++ CHANGELOG.md
@@ -2,6 +2,7 @@ QuantumAnnealing.jl Change Log
 ==============================
 
 ### Staged
+- Add variant of `solve_de` with adaptive solve tolerance
 - Add tools for working with classical Ising models (#8)
 - Add data processing tools (#6)